Added get and set methods for object's VAO's. Switched "window::getWidth" and "window::getHeight" to floats.

+++ TDML/window.cpp
@@ -18,12 +18,12 @@ namespace TDML
 		if(fullscreen) glutFullScreen();
 	}
 
-	int window::getWidth()
+	float window::getWidth()
 	{
 		return width;
 	}
 
-	int window::getHeight()
+	float window::getHeight()
 	{
 		return height;
 	}
